Paginate CLI prefix listing output
Change the way the CLI works so that we fetch small chunks of the result
and output iteratively instead of fetching everyting in a big chunk.

Gives more instant feedback to the user that the result is being
fetched.

In addition, this fixes a bug where result sets over 1200 prefixes would
not be displayed (max_result was set to 1200).

Part of #278