Fix prefix hiding code
Prefix hiding is used to compress the resulting prefix list in the web
interface. If a large assignment (say a /24) contains just one or a few
hosts matching the search query, the result could look bloated. All
non-matching hosts are compressed into a container which are then
hidden so the output is less bloated. If there is only a few prefixes in
a hidden container, we should not hide it though, it just looks silly if
it's one prefix. This fixes that.

Fixes #228