pynipap - Pylint; Line too long
Cut of a few lines that were too long. There's a bunch of lines that are
still too long, but they are typically exception lines with a message
and I don't see that the code is more elegant with them cut off.

Pylint score was originally 7.28 and is now up to 7.34

Part of #274