Bug fix in pynipap for prefix_remove
We rely on getting an exception if we try to fetch an element in the
array that doesn't exist. The exception raised however is not KeyError
but IndexError. IndexError is for lists while KeyErrors are for dicts.
This is a list, thus IndexError is correct!

Don't think we have a issues registered for this.