Fix parent_prefix!
This is almost the final fix. Now parent_prefix is the id of a prefix
from which we will base the search. Only prefixes that are one indent
level below and inside the parent_prefix are included.

A new exception is thrown if the prefix id specified does not exist. We
need to set a good error number for that and handle it in pynipap (and
jnipap).

Part of #206