Add implied VRF to pool in CLI listing
It is now possible to search and list the implied VRF through the CLI
client. This required a few other changes and I took the chance to
rewrite the pool listing to use a smart search instead of the old lame
way.

I really just did the CLI to test. The issue only involved fixing the
backend and possibly pynipap... hmm, this leaves jnipap, right? No one
likes Java anyway.

Fixes #291