Update pool VRF consistency check error message
The mapping of which prefixes belong to which pool is through the
pool_id column in the prefix table. As the database trigger that handles
the consistency is on the prefix table, the given error message was
oriented from perspective that the change is on a prefix - which is true
in the database. The user however, probably sees this from the
perspective of modifying a pool.

This new error message is from a slightly more neutral perspective. It
makes more sense regardless if you think you are editing a prefix or
expanding a pool (ie, modifying the pool).

Fixes #253