Fix bfcache handling
This now uses the 'pageshow' event to trigger things when a page loads
from the bfcache (back forward cache) in some browers (Firefox for
example).

Since jQuery 1.7 is required to register the event, jQuery has been
upgraded to the latest version (1.9.1). No other issues seem to have
come from that. Running the jQuery migration script shows that our
version of jQuery UI relies on a function (attrFn) that is deprecated in
jQuery 1.9.1. We have a custom UI, so I'm somewhat uncertain which
parameters need to be sent to get a new one with the same custom
settings.

Fixes #235