Fix tiptip for jQuery 1.9
tipTip breaks on jQuery 1.9 and since we recently upgraded to that
version, tipTip was broken. There is a fix for this which is currently
in a pull request on tipTips github
(https://github.com/drewwilson/TipTip/pull/15), which I've voted up, but
until it is released, we choose to implement the change ourselves.