Add name and description to VRF selector
This adds the name and description of a VRF to the VRF selector. It's a
bit more code than I thought, mostly due to the click function being a
PITA to handle as it doesn't get enough information from the click event
but needs to traverse a part of the DOM tree.

I thought it worked fine in my first iteration of code, but depending on
if the user clicked the a href or a div (and how deeply nested that div
was), we needed to traverse the tree varyingly number of steps up to the
parent to find the 'data-vrf_id' attribute. It's now performed through a
loop which works well. Another option is to store the data-vrf_id
attribute on all elements of each VRF selector entry, but that's fugly
in my world.

The XHR functions needed to be extended to store the description too.

Fixes #270.