Don't do implicit VRF search for empty string
If we are not on the prefix list page, don't try to do a prefix search.

If the query string is empty, don't try to do a search either.

Also let explicit = null to the prefix search function mean that it
should use the current value from the URI.

Fixes #305 and #299.