Allow changing a prefix to a bigger one
Growing an assignment from say a /27 to a /26 wouldn't work as one of
the consistency check makes sure that the new assignment only contains
hosts. The trigger runs before the update so it finds the old prefix
(/27) being inside the new /26 and since the old prefix is an assignment
the fail checks.

We avoid this little quirk by simply not including the old prefix in the
SELECT.

Fixes #261.