Fix add prefix radio button magic
In from-prefix assignment mode and when a prefix of type assignment has
been selected, the only choice for the new child prefix will be to be of
type 'host'. The radio button for 'reservation' and 'assignment' was
correctly greyed out but the 'host' radio button was not checked. This
fixes that!

Fixes #310.