Change nipap.conf defaults for DB connection
The database connection parameters were previously commented out. The
idea was that the system administrator installing NIPAP would need to go
and change them to get the system up and running. Since things like
passwords should be set to something secure this would be a required
step regardless of other things.

For Travis-CI however, we need a nipap.conf with some sane settings to
just get the system up and running and so I've removed the comments and
set the default password to 'papin'. I would still argue that this is
secure enough since the system administrator would have to explicitly
create his database with the password 'papin' for it to be /insecure/.
The SQL Makefile defaults to creating a random password so unless
someone does PGPASSWORD=papin (which we do in the Travis-CI
installation) then this is a non-issue.

Part of #176.