Fixed VRF selector menu for default VRF
The VRF selector now shows the default VRF per default as well as on
empty searches, just like the old one did. The difference is in that the
old way was to display static information while the new way is to do a
search towards the backend to get the default VRF information. We need
to do that as the name and description of the default VRF can be
changed.

Fixes #292.