Fix Unicode search
The backend broke down on searches containing unicode characters as the
_is_ipv4 function tried to mangle the data through inet_aton which
failed miserably on Unicode and raised an exception which we did not
catch. We catch it now, so all is good.

Also removed some encode/decode mangling which is not needed in Python
2.7.3.

As far as I can tell, unicode searches now works on a recent Python
(2.7.3).

Fixes #247.