Allow editing of the prefix itself via CLI
This opens up the same functionality that pynipap and the web UI already
offers in editing the prefix itself. For example, one could "expand"
1.0.0.8/29 to 1.0.0.0/28 by using this command.

It still comes with a number of limitations imposed by the strict
database constraints we have in place for keeping integrity with regard
to hosts/assignments and so forth.

Fixes #307.