Reset prefix add page on page show
When reloading or going back / forward to the prefix add page, some
things are cached, while others are computed by javascript. The radio
button for selecting prefix allocation method was cached by the bfcache
in certain browsers but the action associated with choosing one radio
button is performed by javascript and upon showing the page from bfcache
that action was not triggered so the page ended up in a non-consistent
state.

This change resets the radio button for prefix allocation method as to
avoid the inconsistent state. It is not the fix that we first came up
with (which was to set onunload="") in the issue, but this works better.
Hope it solves all the cases we hoped to solve.

In the future, we will build upon the page show event for a more
complete state machine, tracking current state in the URL. That work is
documented in #332.

Fixes #311.