Updated jnipap for API change
jnipap now supports the API change where a full object is returned after
edit_* and add_* API calls. In preparation for unit tests (as well as
for improved functionality) the objects were extended with hashCode and
equals-methods.

Lastly, support for the implied_vrf was added to the pool objects.

Fixes #254