Fix db constraints for country value
This adds constraints for making sure the value in the country column is
a two letter alpha character. It is some form of semi-check to verify we
have a ISO 3166-1 compliant value. Doing a proper check towards 3166-1
means we would have to include a list of country codes. We don't want
that as it would likely just become outdated, thus this just performs
the most basic of compliant checks.

Fixes #325.