Fix offset calculation for floating prefix header
The floating prefix header is usually hidden and becomes visible when
the user has scrolled down far enough that the fixed prefix header
is outside of the screen. To achieve this, we get the position of the
fixed header relative to the top and then attach an event handler to
calculate current scroll offset from top.

Problem was that we only retrieved the fixed headers offset from the
screen at page load, and as result is rendered, this could change. Most
notably, the search interpretation will make the search "header" grow a
bit which will change the offset.

This is now fixed by retrieving the position of the fixed header every
time the event handler is called.

Fixes #335.