Add CLI input validation for extraneous arguments
If extraneous arguments are sent to the command parser an exception is
now raised. I hope this doesn't break anything else. The unittest is so
far from complete when it comes to command parsing that it won't help.
I've done some manual testing which didn't show anything else being
broken.

Fixes #276.