Add version info to nipap CLI
This adds a simple --version parameter to the nipap CLI client to list
the version information of the CLI and the pynipap library.

The change is slightly larger than I intended, but the __version__ info
is defined in the __init__.py file and as nipap_cli was originally
imported with:
 from nipap_cli import nipap_cli
we could not reach the __version__ variable. Thus, we now import
nipap_cli and nipap_cli.nipap_cli instead.

Using optparse is likely overkill, but I figured; why not?

Part of #339.