Renamed 'nipap' module to 'backend'
The module 'nipap' used by nipapd was renamed to 'backend' due to the
unfortunate naming collision which prevented modules in the nipap
package to access the package as the nipap module had precedence over
the package.

Also, after all this messin' around, added an API function to fetch the
nipapd version.

part of #339