Add debian configuration scripts
This will have debconf ask the user questions to configure / setup
nipapd. At this point, two actions are supported;
 * if nipapd should be started upon system startup
 * Automatically configuring postgresql for NIPAP

In the future the idea is that this script will help with upgrading the
database structure and similar things as well.

Everything to keep installation as smooth as possible.

Part of #122.