Fix dist config file
We got these funny log messages from nipapd:
2013-05-10 16:39:29,793: INFO     Listening to address 127.0.0.1
; 127.0.0.1. Use 0.0.0.0 to listen on all
; addresses (change requires restart) on port 1337

The configparser lib parses lines starting with a whitespace as a
multi-line value to the previous key. In the default config file there
was a multi-line comment, but configparser though it was a value.

Default config file has been updated to avoid this.

Fixes #346.