Update debian control files
Update nipapd dependancy to 2.7, since that is what we need for unicode
support and it is included in debian 7.0. I don't see a reason why to
depend on something older.

Add dependancy on postgresql-9.1-ip4r. I've asked the debian maintainer
for ip4r to att a meta-package called postgresql-ip4r so that we don't
have to depend on 9.1 today and postgresql-9.2-ip4r tomorrow... we'll
see how that goes, until then we will depend on the 9.1 version of the
package.

Last but not least we depend on the package sudo. This is a tad silly,
but in the postinst script of nipapd, sudo is used extensively to load
things into PostgreSQL. Perhaps that could be replaced with 'su'
instead, but as it requires some reworking of the code I'll leave that
for later.