Add postinst script to nipap-www
postinst script for nipap-www debian package now asks question whether
it should automatically add a new user to the local auth db for the web
interface and configure the new user in the nipap.conf configuration
file as the XMLRPC-URI.

This is quite tricky and a bit of a kludge as nipap-www does not depend
on the nipapd package and this auto-setup help is only possible if
nipapd is installed on the same machine. Since we need to be certain the
nipapd package files are unpacked, the question is not raised in the
'config' file as is customary but in the postinst file.

Also, the configuration script for nipapd needs to be run before all of
this is run and since we cannot make a hard dependancy on nipapd
(because one should be able to run it on separate machines), we cannot
possible know this. Since nipap-www won't break if nipapd is not
installed, I'm saying this solution is good enough for the moment. Best
case is you get help with adding a user for nipap-www while the worst
case is that nothing happens (you can then get the assistance by doing
dpkg-reconfigure nipap-www).

For the future, we should figure out a better way of handling all this,
possibly telling apt the order that packages should be installed in but
without having any hard dependencies - is that possible?

Fixes #345