Raise max results for vrf listing
Tab completion for vrf listings were limited by the default max_result
of 50. The max_result is now set to 100000 for these queries and so
we've probably avoided the issue for the time being.. not sure what
would be the optimal solution. Tab completion does give much room for
things like error/warning messages and so just raising the limit feels
like the best solution for the time being. Few should need to go above
100000. Listing 10k VRFs takes about a second on my laptop..

Fixes #353.