Run nipapd as nipap user
Change default for nipap configuration file so as to run nipapd as the
user and group 'nipap'.

Since we now run nipapd as a non-root user we cannot create the
nipapd.pid file in /var/run due to insufficient privileges. Instead a
new directory, /var/run/nipap, is created at install and is owned by the
nipap user. The pid file is written to that directory.

Running nipapd as the user nipap allows us to use connect to PostgreSQL
using a UNIX socket instead of over TCP. This will make the installation
of new NIPAP systems easier as there is no need to edit the
PostgreSQL configuration file.

Fix postrm script to remove nipap user and group if nipapd is removed.

For new installations nipapd will now be run as nipap and the database
authentication will be using UNIX socket. Configuration file will not be
changed for old installations (ie upgrades), but please see the
nipap.conf.dist for an example.

Fixes #361.