Fix permissions on local_auth.db
I commited a change that made the nipapd.postinst script set the
permissions on the local_auth.db file to 0640 and ownership to
root:nipap. This works fine on a system just running nipapd but on a
system which also hosts the web UI it will break things as the web UI
needs to be able to read the local_auth.db file as well. The web UI is
typically run by apache2 which in turn runs as www-data and subsequently
does not have read access to local_auth.db. I want to add that I did try
my changes on the nipap-demo but due to my login being cached I didn't
run into this problem.

Long term fix would probably be to implement #207. Fix for now is merely
to allow everyone to read the file again.