Fix pid file issues in init script
The pid file resides in /var/run/nipap but as /var/run is symlinked to
/run which in turn is a tmpfs, the /var/run/nipap directory is not
retained through reboots.

Previously the postinst script was responsible for creating the
/var/run/nipap directory but as we need to do this at least on every
boot, the functionality has now been moved to the nipapd init script. In
addition some of the functions have been made more robust to handle the
lack of a /var/run/nipap directory.

Last but not least, the postinst script now starts nipapd at the end of
the installation!