PoC backend on Flask-XMLRPC and Tornado
This implements a new backend engine for NIPAP that uses the Flask
XML-RPC framework instead of Twisted. Tornado is used to serve Flask
requests and allows pre-forking to handle multiple requests in parallel.

Authentication is performed similarly to how it was done with Twisted.
The request is mangled and a auth-object, produced by our AuthFactory,
is inserted.

The XML-RPC echo function now support sleeping for a specified amount of
time so that it becomes easier to test parallel requests.

xmlbench.py implements a small test benchmark that sends a bunch of
requests in parallel to the nipap XML-RPC API and measures the time
taken to respond. While not a complete and finished product in itself,
it comes in fairly handy for testing parallel requests.

Fixes #273.