Fix bug in helper-nipap
This makes helper-nipap a bit more tolerant to errors in that it will
find the first argument called 'nipap' and continue processing all
arguments after that first argument. This means that helper-nipap can be
called like "helper-nipap nipap address" or "helper-nipap address" and
it will still produce the same output.

Fixes #375.