Differentiate between pageload and popstate
Google Chrome and Firefox differs in how the history API handles a page
load. Firefox emits a 'pageload' event, while Chrome emits both a
'pageload' and a 'popstate' event. Added code from
http://stackoverflow.com/questions/6421769/popstate-on-pages-load-in-chrome
which tries to differentiate between the popstate event which is emitted
by Chrome when first entering a page to the one which is emitted when
the user uses the back/forward buttons of the browser.

So far it catches most occurencies, but not all, which results in the
page being loaded twice and a first access to the prefix list page
performs an explicit search with an empty search string.

Part of #208