Removed performPrefixSearch from pageshow handler
...as it'll be called when the VRF list has been returned. Also modified
when the setSearchPrefixURI is called from performPrefixSearch.

This seems to strike the correct balance on when to update URI and when
to perform search. Apparantly using the back button to jump from another
page to the prefix list does not trigger the popstate event in Firefox,
but now that seems to be handled correctly...

Fixes #208 \o/