Improve tab completion for args with escapes
Tab completion was broken for certain commands where the object
contained escaped whitespace, quotes or similar. For example, the
command 'nipap pool resize foo\ bar remove <tab>' would not work.
helper-nipap needs to read the name of the pool so that it can complete
the member prefixes based on that information. bash however, passed the
literal 'foo\ bar' instead of 'foo bar' which broke helper-nipap.

That is fixed by mangling everything through shlex before to remove
escaped whitespace, quotes and similar.

Kind of part of #375.