Let Travis build deb from source
Let Travis build deb packages from source and install those instead of
pulling debs from the official NIPAP repo. This tests that the package
building works as well as the installation scripts!

Modification done to the database will also be included in the setup
instead of the potentially old database structure that the stable NIPAP
build ships with.