Add VLAN attribute to prefixes
This adds a new column in the database (simply called 'vlan') and
exposes the value of that column over XML-RPC. Pynipap, the CLI and the
web UI have been updated to display the new attribute and allow editing
of it.

There are currently no constraints other than that the value must be an
integer.

Fixes #393