Fix mistakes from implementation of 190 and 393
I completely missed adding VLAN and Tags to the prefix add form, as that
is a separated form from prefix edit. That has now been fixed and it's
possible to specify VLAN and tags when adding a prefix.

There was also a bug in where an empty VLAN string ('') would not be
accepted as the database only accepts integers for the VLAN column, not
an empty string. This has been fixed, both for editing and adding a
prefix.

We should try to merge the two forms to make subsequent additions of new
attributes easier.

This fixes #190 and #393.