List top level prefixes on empty search
When performing an "empty search", ie with a zero-length search string,
only list the top level prefixes (indent = 0). An empty search is
probably a sign of that the user wants to see everything and showing
only top level prefixes is probably a better starting point to give an
overview than directly listing and expanding all prefixes.

Fixes #340.