Add new timestamp values to prefixes
The backend now serves two new values, the 'added' and 'last_modified'
columns of the database which are the respective timestamps for when a
prefix was first added and last modified. Both values are also
searchable via the prefix smart search function. Neither value is
writable as the backend updates them automatically.

xmlrpclib, both in the backend and in pynipap, is used with
use_datetime=True which means that the timestamps will automatically be
converted into datetime objects for easier use.

Pynipap does naturally not support saving the values as they are
read-only. The CLI will show the 'added' and 'last_modified' timestamps
when viewing a prefix. At this time, they are mostly meant for internal
usage or via the API, e.g. to bring up all prefixes that have been
changed the last hour to sync with another system.

Fixes #367.