Fix rendering of inherited tags
When a prefix has a tag, a small icon is rendered next to the prefix and
onMouseOver a tooltip dialog will list the tags and inherited tags.
First, the icon and tooltip was not available if the prefix only had
inherited tags (ie, no tags). Secondly, when the tooltip was displayed,
the list contained the array index and not the name of the inherited
tag. Both issues have been resolved.

Fixes #430.