Add option to include neighbors in prefix search
Neighbors to a matching hosts were previously always included in the
search result, regardless of the value of parents_depth and
children_depth. With this fix, neighboring hosts to a matching host is
only included when the include_neighbors option is set.

The web UI should include neighbors, so it has been modified to set
include_neighbors to True.

Also added a new test suite for testing the logic behind parents_depth /
children_depth and the new include_neighbors.

Fixes of #416.