Add Makefile target for automatic version bump
All the Makefiles for the different NIPAP components now have a new
target called 'bumpversion'. It will bump up the version based on the
content of the NEWS file in the project root. sed and the new
news2dch.py script is used to update individual files containing version
information and the debian/changelog respectively.

Fixes #435. This actually accomplishes more than we originally set out
to do with #435 but I guess that is ok ;)