Auto-update debian config & postinst script
Makefile target 'bumpversion' will now update the database version in
the debian package scripts 'config' and 'postinst' so that these are
aligned with what is in the nipapd database file. The nipapd.config
script has been out of sync which caused the prompt to update the
database to be displayed when it is actually not needed.

Fixes #424.