Refactor xmlrpc to allow import without init
The xmlrpc module initiated an instance of the nipap backend and a few
other things which made it impossible to simply import the module
without first setting up all the prerequisites. Unittesting and similar
was made difficult.

This change moves everything into a class so that nothing is
instantiated upon import and thus unittesting or things like generating
docs (by sphinx) is easier.