Sphinx changes for readthedocs.org
readthedocs (RTD) automatically builds our documentation. Modules are
imported which means all dependencies needs to be present but as RTD
fails to build things like python-ldap, we instead mock that module.
Other modules, such as Flask, have proven to be really difficult to mock
and they are instead installed.

Part of #283.