Update installation guide for Debian (derivatives)
I don't really know how this happened, but the guide on the web page
(stored in a .html on gh-pages branch) is way ahead of this .rst file.
Have we updated the documentation just in the HTML and not in the source
file???

Weird. Anyway, I've backported the changes from the .html file to this
.rst file so they are now in sync again.