Remove 'guest' user from default shipping
We shouldn't ship any default user in the local auth database - it's
better to let the user add one himself.

Given that nipapd only listens on 127.0.0.1 per default it's not a very
big security issue but nonetheless it could raise some concerns ;)

.travis.yml is updated to add a user that can be used for testing and
all unittests are updated to use this new user.

Fixes #446.