Increase minimum column width for VRF in CLI
VRF RT column width wasn't set so when listing multiple VRFs it would be
fairly common with some mis-alignment between prefixes in different
VRFs. Now the minimum width is set to 14 which means that it looks fine
for most VRF RTs. Long RTs, like with 4-byte ASn, will still result in
some alignment issues, but this should fix the majority of cases.

Fixes #454.