Introduced AngularJS by converting VRF list page
Began introducing AngularJS into the NIPAP web UI by converting the
/vrf/list page. The implementation is currently a mixture between
old-style and new-Angular-style code; for example confirmation dialogs
are not yet "Angularified".

An Angular module for the entire NIPAP page was added and controllers
will be added for each page. A VRFListController was created for the
/vrf/list page.

As we're heading towards a web site which simply can't be used without a
JavaScript-enabled browser, I chose to remove things such as
href-attributes on anhors which also have click event handlers in
JavaScript and will continue down that road.

Fixes #347 and fixes #338