Add functions to create & upgrade sqlite database
The SqliteAuth class  has new functions to check that the database
schema is of the latest "version" and to upgrade the database.

nipapd is now performing pre-start checks to make sure that the sqlite
database is of the latest version, which is required for correct
operation. If not, it will exit and print an error message.

nipap-passwd has new arguments to check if the db is of latest version,
to create a new database and upgrade an existing database.

Part of #438.