Fix debian config script
We cannot use db_reset to reset the answer to a possibly already
answered question as that breaks the use of debconf-communicate.

The use of db_reset remains for the upgrade questions - I see no good
way of solving those right now. I did look into how the config script is
invoked, if we could possibly figure something out there, but we do not
get any hints.