Set stdout to always output UTF8
Outputting UTF-8 usually works fine as Python recognizes the terminal as
UTF-8 compatible or in the case of ISO-8859-1 it at least has a way of
reencoding data so that characters like å, ä, ö can be displayed
properly.

However, when piping the output, Python no longer knows what the
receiving end will accept and so it defaults to US-ASCII or something
similarly ancient.

This sets the default output to UTF-8 when Python is unable to determine
the output encoding, typically when stdout is redirected. Hope the
other end can receive it!

Fixes #491.