Add basic statistics support
This introduces basic statistics support in NIPAP for prefixes and adds
preparatory changes for pools and VRFs.  Do note how there is no time dimension
to this just yet, there's only statistics for number of used addresses and so
forth, adding a time dimension would come as a second step.

New columns in db for VRFs (ip_net_vrf), pools (ip_net_pool) and
prefixes (ip_net_plan) containing statistics. Most statistics values are
stored as numeric(40) which allows for very large numbers (like 2^128
for IPv6).  The values for VRFs are;
 - total number of IPv4 prefixes in VRF
 - total number of IPv6 prefixes in VRF
 - total number of IPv4 addresses in VRF
 - total number of IPv6 addresses in VRF
 - number of used IPv4 addresses in VRF
 - number of used IPv6 addresses in VRF
 - number of free IPv4 addresses in VRF
 - number of free IPv6 addresses in VRF

For pools:
 - total number of member IPv4 prefixes in pool
 - total number of member IPv6 prefixes in pool
 - number of IPv4 prefixes assigned from pool
 - number of IPv6 prefixes assigned from pool
 - total number of IPv4 addresses in pool
 - total number of IPv6 addresses in pool
 - number of used IPv4 addresses in pool
 - number of used IPv6 addresses in pool
 - number of free IPv4 addresses in pool
 - number of free IPv6 addresses in pool

For prefixes we have;
 - total number of addresses
 - number of used addresses
 - number of free addresses

All statistics for prefixes are for direct children, ie we will not show
if there are free addresses further down a hierarchy of prefixes.

Database triggers automatically update statistics for prefixes on
INSERT/UPDATE/DELETE operations so that statistics are always kept up to date.

There's quite a few different triggers associated with our tables, in
particular ip_net_plan, to perform various checks and calculations of
things like indent, inherited tags, number of children and now
statistics. It's a bit of a mess. The idea behind having lots of
different triggers and trigger functions was that they would each be
kept simple and thus easily maintanable. The issue is that there are
dependencies between the functions, so now we have to assure they are
being run in the right order. For this reason I'm now trying a different
approach of joining more features into fewer trigger functions but still
keeping each feature as a separate section within the function,
hopefully making it still rather easy to maintain.

This change moves indent and children calculation from its own trigger
function into a shared function where the new statistics are handled as well.

Naturally, there's a new comprehensive test suite for statistics.

nipapd will serve these new values and the CLI can display them.

NOTE: Most statistics values are sent over XML-RPC as strings as the XML-RPC
standard doesn't specify a datatype for anything larger than an integer.
Internally in the backend module, we still deal with it as a long - the
actual conversion is done in xmlrpc.py so if we ever add a second API
(JSON-RPC anyone?) then we could still pass it as a number, if that API
would support it. Pynipap will convert it back into a long on the client side
so clients using pynipap will not have to deal with this peculiarity.

Part of #248.